Teach gh-activity-importer to clone remote repos
Currently it clones it to a local directory, 'dest', and then does
the same thing as before, but in the future it can clone to an in-
memory tree, apply the commits then automatically push the changes
to the single remote.

Also add more configurable logging. This isn't exposed in the CLI,
but the app consistently uses the logger.